Fix small selection bug in IE8
When filling in an input box via jQuery/javascript that is currently
focused `getSelectionStart` returns -1 which causes the first character
of the new data in the input box to become selected. This patch does not
allow the function to return less than 0 which fixes this somewhat
superficial bug. This was only a bug in IE8; IE7 and IE9 worked as
expected.